break up big passive components
too big to refactor
