+++ electronics_abstract_parts/Categories.py
@@ -1,21 +1,5 @@
 from electronics_model import *
-from itertools import chain
 
-
-@abstract_block
-class E24SeriesDevice(Block):
-  """Non-physical "E24 series device" used to affect parameters."""
-
-  E24_SERIES = [
-    [1.0],  # E1
-    [2.2, 4.7],  # E3
-    [1.5, 3.3, 6.8],  # E6
-    [1.2, 1.8, 2.7, 3.9, 5.6, 8.2],  # E12
-    [1.1, 1.3, 1.6, 2.0, 2.4, 3.0, 3.6, 4.3, 5.1, 6.2, 7.5, 9.1],  # E24
-  ]
-
-  # Zigzag (lower E-series first) of the E12 series
-  E24_SERIES_ZIGZAG = list(chain(*E24_SERIES))
 
 @abstract_block
 class DummyDevice(Block):
@@ -42,7 +26,7 @@ class DiscreteSemiconductor(DiscreteComponent):
 
 
 @abstract_block
-class PassiveComponent(DiscreteComponent, E24SeriesDevice):
+class PassiveComponent(DiscreteComponent):
   """Passives components, typically used as part of an application circuit."""
   pass
