+++ electronics_abstract_parts/ESeriesUtil.py
@@ -1,0 +1,58 @@
+from typing import Sequence, Optional
+from electronics_model import *
+from itertools import chain
+import math
+
+
+class ESeriesUtil:
+  """Helper methods for working with the E series of preferred numbers."""
+  @staticmethod
+  def zigzag_range(start: int, end: int) -> Sequence[int]:
+    if start >= end:
+      return []
+
+    center = int((start + end - 1) / 2)
+    lower = list(range(start, center))
+    upper = list(range(center + 1, end))
+    output = [center]
+
+    while lower or upper:
+      if lower:
+        output.append(lower.pop(0))
+      if upper:
+        output.append(upper.pop(0))
+
+    return output
+
+  @staticmethod
+  def round_sig(x: float, sig: int) -> float:
+    return round(x, sig-int(math.floor(math.log10(abs(x))))-1)
+
+  @classmethod
+  def choose_preferred_number(cls, range: Range, tolerance: float, series: Sequence[float], sig: int) -> \
+      Optional[float]:
+    lower_pow10 = math.floor(math.log10(range.lower))
+    upper_pow10 = math.ceil(math.log10(range.upper))
+
+    powers = cls.zigzag_range(lower_pow10, upper_pow10)  # prefer the center power first, then zigzag away from it
+    # TODO given the tolerance we can actually bound this further
+
+    for value in series:
+      for power in powers:
+        pow10_mult = math.pow(10, power)
+        value_mult = cls.round_sig(value * pow10_mult, sig)  # this prevents floating point weirdness, eg 819.999
+        if Range.from_tolerance(value_mult, tolerance) in range:
+          return value_mult
+
+    return None
+
+  E24_SERIES = [
+    [1.0],  # E1
+    [2.2, 4.7],  # E3
+    [1.5, 3.3, 6.8],  # E6
+    [1.2, 1.8, 2.7, 3.9, 5.6, 8.2],  # E12
+    [1.1, 1.3, 1.6, 2.0, 2.4, 3.0, 3.6, 4.3, 5.1, 6.2, 7.5, 9.1],  # E24
+  ]
+
+  # Zigzag (lower E-series first) of the E12 series
+  E24_SERIES_ZIGZAG = list(chain(*E24_SERIES))
