+++ electronics_lib/test_capacitor.py
@@ -1,10 +1,6 @@
 import unittest
 
-from electronics_abstract_parts import *
-import electronics_abstract_parts
-import electronics_model
-from . import *
-from . import Passives
+from .PassiveCapacitor import SmtCeramicCapacitor
 from .test_passive_common import *
